7th commit - V0.75
Fixed some error & function from the client's request
1. I made it to adjust the volume max automatically. Please note that
the volume will vary depending on devices or earphones. It is impossible
to bypass devices' volume control.
2. I corrected. Please let me know if there is any other typo or
grammatical error. It would be better if you write and send me the
instruction you want to display.
3. I increased the font size to 0.15 inch at trial page and screening
page. I will modify the other pages if you like it. But please note that
it is not a good idea to specify the font size because it depends on
device's screen size. If you want to fit in a particular device(such as
Galaxy S3), I will try to fit to it.
4. It is nothing but shuffled randomly. I carefully guess the reason you
think is the test set has only 7 different cases each ear two times(i.e.
7 x 2 x 2 = 28 times). It is possible that the test cases are ordered
same frequency together high frequently. At least in my test it is
shuffled well.
5. I found no problem to hear each test set. I can hear all frequency
and decibel sets(125/40, 250/40, 500/40, 1000/50, 2000/65, 4000/80 and
8000/85). You can see the frequency and decibel result data after
releasing your finger from the screen with previous version. Please tell
me which frequency/dB you cannot hear and the name of the device which
you are using.
6. I removed it. Don't worry about that it was my debugging message.
7. I removed two steps in Trial page and Screening page. Please check
it.

8. I want to make sure what is different in user information between
Screener and Subject. In "Add New Screener" you stated Name, Username
and Password, and in "Subject Information" you stated First Name, Last
Name and User ID(Please see the flow chart at page 2-3 in Hearing App
Design Layout.docx). It is confusing how would you like to manage them
and if they share same contents(such as Login and Screening Test
subject) or not.

9. It seems to be hard because I used someone's library to show the
graph. They don't provide inverting y-axis function. I will keep
researching if I can do it or not.
10. I fixed it. I temporarily fixed the user name as my name. I will
finish it soon after clarification of the user information.